backend/requirements.txt
telethon>=1.37.0
tqdm>=4.66.0
colorama>=0.4.6
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
websockets>=11.0